Move `JobStatusType` from models to schemas (ref #935)
