+++ fractal_server/app/models/job.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from enum import Enum
 from typing import Any
 from typing import Optional
 
@@ -10,32 +9,8 @@ from sqlmodel import Field
 from sqlmodel import SQLModel
 
 from ...utils import get_timestamp
+from ..schemas import JobStatusType
 from ..schemas.applyworkflow import _ApplyWorkflowBase
-
-
-class JobStatusType(str, Enum):
-    """
-    Define the job status available
-
-    Attributes:
-        SUBMITTED:
-            The workflow has been applied but not yet scheduled with an
-            executor. In this phase, due diligence takes place, such as
-            creating working directory, assemblying arguments, etc.
-        RUNNING:
-            The workflow was scheduled with an executor. Note that it might not
-            yet be running within the executor, e.g., jobs could still be
-            pending within a SLURM executor.
-        DONE:
-            The workflow was applied successfully
-        FAILED:
-            The workflow terminated with an error.
-    """
-
-    SUBMITTED = "submitted"
-    RUNNING = "running"
-    DONE = "done"
-    FAILED = "failed"
 
 
 class ApplyWorkflow(_ApplyWorkflowBase, SQLModel, table=True):
